+++ core/renderer.h
@@ -389,7 +389,7 @@ public:
 		static	int						bucketWidth,bucketHeight;						// Bucket dimentions in samples
 		static	int						netXBuckets,netYBuckets;						// The meta bucket size
 		static	int						threadStride;									// The number of buckets per thread at a time
-		static	int						geoCacheSize;									// The ammount of memory to dedicate to tesselation caches
+		static	size_t					geoCacheSize;									// The ammount of memory to dedicate to tesselation caches
 		static	int						maxEyeSplits;									// Maximum number of eye splits
 		static	float					tsmThreshold;									// Transparency shadow map threshold
 		static	float					dimgThreshold;									// Transparency shadow map threshol
